MAINT: Refactor data model structure
