+++ schema/definitions/0.8.0/schema/fmu_meta.json
@@ -430,7 +430,7 @@
           "title": "metadata_class"
         },
         "fmu": {
-          "$ref": "#/$defs/FMUCaseAttributes"
+          "$ref": "#/$defs/FMUBase"
         },
         "masterdata": {
           "$ref": "#/$defs/Masterdata"
@@ -955,7 +955,7 @@
       "title": "DomainReference",
       "type": "string"
     },
-    "FMUAttributes": {
+    "FMU": {
       "dependencies": {
         "aggregation": {
           "not": {
@@ -1033,10 +1033,10 @@
         "model",
         "context"
       ],
-      "title": "FMUAttributes",
+      "title": "FMU",
       "type": "object"
     },
-    "FMUCaseAttributes": {
+    "FMUBase": {
       "description": "The ``fmu`` block contains all attributes specific to FMU. The idea is that the FMU\nresults data model can be applied to data from *other* sources - in which the\nfmu-specific stuff may not make sense or be applicable.",
       "properties": {
         "case": {
@@ -1050,7 +1050,7 @@
         "case",
         "model"
       ],
-      "title": "FMUCaseAttributes",
+      "title": "FMUBase",
       "type": "object"
     },
     "FMUContext": {
@@ -4224,7 +4224,7 @@
           "$ref": "#/$defs/File"
         },
         "fmu": {
-          "$ref": "#/$defs/FMUAttributes"
+          "$ref": "#/$defs/FMU"
         },
         "masterdata": {
           "$ref": "#/$defs/Masterdata"
