+++ src/fmu/dataio/_model/__init__.py
@@ -1,4 +1,4 @@
-from .meta import Root, dump
+from .root import Root, dump
 
 __all__ = [
     "dump",
