+++ src/fmu/dataio/_model/enums.py
