+++ src/fmu/dataio/_model/specification.py
