--- src/fmu/dataio/datastructure/_internal/__init__.py
@@ -1,7 +1,0 @@
-"""
-This module, `datastructure._internal`, contains internal data structures that
-are designed to depend on external modules, but not the other way around.
-This design ensures modularity and flexibility, allowing external modules
-to be potentially separated into their own repositories without dependencies
-on the internals.
-"""
