+++ src/fmu/dataio/providers/objectdata/_faultroom.py
@@ -5,9 +5,10 @@ from typing import TYPE_CHECKING, Final
 
 from fmu.dataio._definitions import ExportFolder, ValidFormats
 from fmu.dataio._logging import null_logger
-from fmu.dataio.datastructure.meta.content import BoundingBox3D
-from fmu.dataio.datastructure.meta.enums import FMUClass, Layout
-from fmu.dataio.datastructure.meta.specification import FaultRoomSurfaceSpecification
+from fmu.dataio._model.data import BoundingBox3D
+from fmu.dataio._model.enums import FMUClass, Layout
+from fmu.dataio._model.specification import FaultRoomSurfaceSpecification
+from fmu.dataio.readers import FaultRoomSurface
 
 from ._base import (
     ObjectDataProvider,
