+++ src/fmu/dataio/providers/objectdata/_provider.py
@@ -94,7 +94,7 @@ import xtgeo
 
 from fmu.dataio._definitions import ExportFolder, ValidFormats
 from fmu.dataio._logging import null_logger
-from fmu.dataio.datastructure.meta.enums import FMUClass, Layout
+from fmu.dataio._model.enums import FMUClass, Layout
 from fmu.dataio.readers import FaultRoomSurface
 
 from ._base import (
