+++ tests/test_units/test_dataio.py
@@ -10,9 +10,9 @@ from pathlib import Path
 import pydantic
 import pytest
 import yaml
+from fmu.dataio._model.enums import FMUContext
 from fmu.dataio._utils import prettyprint_dict
 from fmu.dataio.dataio import ExportData, read_metadata
-from fmu.dataio.datastructure.meta.enums import FMUContext
 from fmu.dataio.providers._fmu import FmuEnv
 
 # pylint: disable=no-member
